Issue #22: JSON Schema (#23)
* feat: add a new input field for the form signature for the json_schema

* feat: Update max_token value for gpt-4o model

The `max_token` value for the `gpt-4o` model in the `gpt.py` file was changed to `None`. This change allows for unlimited token length when making API calls with the `gpt-4o` model.

* feat: Update company information in expected.json

The company information in the `expected.json` file was updated to reflect the new details of GreenGrow Inc. This change includes the company name, address, website, and phone number.

* fix: remove newline at end of file in test_inspection.py

* chore: Update test_gpt.py with translated warranty information and nutrient values

* feat: Add translated nutrient values for ingredients in expected.json

The code changes include adding nutrient values for ingredients in the expected.json file. This enhancement improves the accuracy and completeness of the data. The commit message follows the established convention of using a "feat" prefix to indicate a new feature or enhancement.

* feat: Update nutrient values in expected.json

The code changes involve updating the nutrient values in the expected.json file. This improves the accuracy and completeness of the data. The commit message follows the established convention of using a "feat" prefix for new features or enhancements.

* feat: Update nutrient values in expected.json

* refactor: Refactor field validation in inspection.py

Refactor the field validation in the `GuaranteedAnalysis` and `FertilizerInspection` classes in `inspection.py`. The `replace_none_with_empty_list` methods have been updated to use the `field_validator` decorator instead of the `model_validator` decorator. This change improves the readability and maintainability of the code.